review fixes part - 1
