+++ src/models/Author.js
@@ -1,25 +1,31 @@
 import { DataTypes } from "sequelize";
-import { sequelize } from "../utils/database.js";
+import sequelize from "../utils/database.js";
 
-const Author = sequelize.define('Author', {
-    id: {
-        type: DataTypes.UUID,
-        defaultValue: DataTypes.UUIDV4,
-        primaryKey: true,
-        allowNull: false
+const Author = sequelize.define(
+    "Author",
+    {
+        id: {
+            type: DataTypes.UUID,
+            defaultValue: DataTypes.UUIDV4,
+            primaryKey: true,
+            allowNull: false,
+        },
+        name: {
+            type: DataTypes.STRING,
+            allowNull: false,
+        },
+        biography: {
+            type: DataTypes.STRING,
+            allowNull: false,
+        },
+        bornDate: {
+            type: DataTypes.DATE,
+            allowNull: false,
+        },
     },
-    name: {
-        type: DataTypes.STRING,
-        allowNull: false
-    },
-    biography: {
-        type: DataTypes.STRING,
-        allowNull: false
-    },
-    bornDate: {
-        type: DataTypes.DATE,
-        allowNull: false
+    {
+        timestamps: true,
     }
-});
+);
 
 export default Author;
