+++ src/models/Book.js
@@ -1,25 +1,31 @@
 import { DataTypes } from "sequelize";
-import { sequelize } from "../utils/database.js";
+import sequelize from "../utils/database.js";
 
-const Book = sequelize.define('Book', {
-    id: {
-        type: DataTypes.UUID,
-        defaultValue: DataTypes.UUIDV4,
-        primaryKey: true,
-        allowNull: false
+const Book = sequelize.define(
+    "Book",
+    {
+        id: {
+            type: DataTypes.UUID,
+            defaultValue: DataTypes.UUIDV4,
+            primaryKey: true,
+            allowNull: false,
+        },
+        title: {
+            type: DataTypes.STRING,
+            allowNull: false,
+        },
+        description: {
+            type: DataTypes.STRING,
+            allowNull: false,
+        },
+        publishedDate: {
+            type: DataTypes.DATE,
+            allowNull: false,
+        },
     },
-    title: {
-        type: DataTypes.STRING,
-        allowNull: false
-    },
-    description: {
-        type: DataTypes.STRING,
-        allowNull: false
-    },
-    publishedDate: {
-        type: DataTypes.DATE,
-        allowNull: false
+    {
+        timestamps: true,
     }
-});
+);
 
 export default Book;
