+++ src/resolvers/authorResolver.js
@@ -1,0 +1,5 @@
+export const authorResolver = {
+    books: async (parent, args, context) => {
+        return await context.loaders.bookLoader.load(parent.id);
+    },
+};
