+++ src/resolvers/bookResolver.js
@@ -1,0 +1,8 @@
+export const bookResolver = {
+    authors: async (parent, args, context) => {
+        return await context.loaders.authorLoader.load(parent.id);
+    },
+    reviewsAndRating: async (parent, args, context) => {
+        return await context.loaders.reviewLoader.load(parent.id);
+    },
+};
