+++ src/utils/database.js
@@ -2,25 +2,27 @@ import { Sequelize } from "sequelize";
 import { config } from "dotenv";
 config({ path: "./src/config.env" });
 
-const PROD_DB_PATH = process.env.PROD_DB_PATH;
-const ENV = process.env.ENV;
-console.log("Env: ", ENV);
+let DB_PATH = "sqlite::memory:";
 
-let DB_PATH = PROD_DB_PATH;
-
-if (ENV === "development") {
-    DB_PATH = "sqlite::memory:";
-    console.log("Connecting to development DB");
+if (process.env.ENV === "production") {
+    console.log("Connecting to production DB");
+    DB_PATH = process.env.PROD_DB_PATH;
 }
 
-export const sequelize = new Sequelize(DB_PATH, {
+const sequelize = new Sequelize(DB_PATH, {
     dialect: "postgres",
 });
 
-try {
-    await sequelize.authenticate();
-    console.log("Connected to DB");
-} catch (error) {
-    console.log("Error connecting to DB: ", error);
-    process.exit(1);
-}
+const initializeDatabase = async () => {
+    try {
+        await sequelize.authenticate();
+        console.log("Connected to DB");
+    } catch (error) {
+        console.log("Error connecting to DB: ", error);
+        process.exit(1);
+    }
+};
+
+initializeDatabase();
+
+export default sequelize;
